- gemsclaim compatibility issues solved - standalone final mutex replaced with conditional variable - irt_lock_t renamed to irt_mutex_t - useless printfs removed - win spinlock implementation removed from thread.h - some warnings fixed

+++ code/runtime/include/include_gems/stdint.h
@@ -38,7 +38,7 @@
 #ifndef __GUARD_INCLUDE_GEMS_STDINT_H
 #define __GUARD_INCLUDE_GEMS_STDINT_H
 
-typedef long long intptr_t;
-typedef unsigned long long uintptr_t;
+typedef int intptr_t;
+typedef unsigned int uintptr_t;
 
 #endif // ifndef __GUARD_INCLUDE_GEMS_STDINT_H
